requirements.txt
arrow==0.16.0
astroid==2.4.2
attrs==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
html5validator==0.3.3
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
nodeenv==1.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
